Use logical type rather than Original ones. Rename packages according to the group id
