+++ src/test/java/io/github/anthorx/parquet/sql/converter/RecordFieldConverterTest.java
@@ -1,12 +1,26 @@
-package org.arx.parquet.sql.converter;
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
 
+package io.github.anthorx.parquet.sql.converter;
+
+import io.github.anthorx.parquet.sql.model.RecordField;
+import io.github.anthorx.parquet.sql.model.SQLField;
 import org.apache.parquet.io.api.Binary;
-import org.arx.parquet.sql.model.RecordField;
-import org.arx.parquet.sql.model.SQLField;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
-import java.sql.Time;
 import java.sql.Timestamp;
 
 import static org.hamcrest.MatcherAssert.assertThat;
